requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
requests>=2.31
pydantic>=2.6
python-dotenv>=1.0